test: Robustly check the Instruction page titles
